# Fix merkle tree rebuilds