--- a/polycommit/src/marlin_pc/gadgets/marlin_kzg10.rs
+++ b/polycommit/src/marlin_pc/gadgets/marlin_kzg10.rs
@@ -153,7 +153,9 @@ where
             labels.1.insert(label);
         }
 
-        // eprintln!("before PC combining commitments: constraints: {}", cs.num_constraints());
+        if cfg!(debug_assertions) {
+            println!("before PC combining commitments: constraints: {}", cs.num_constraints());
+        }
 
         let zero = PG::G1Gadget::zero(cs.ns(|| format!("g1_zero")))?;
 
@@ -354,7 +356,9 @@ where
             combined_evals.push(combined_eval_reduced);
         }
 
-        // eprintln!("before PC batch check: constraints: {}", cs.num_constraints());
+        if cfg!(debug_assertions) {
+            println!("before PC batch check: constraints: {}", cs.num_constraints());
+        }
 
         // Perform the batch check.
         {
@@ -528,7 +532,9 @@ where
                 &[prepared_beta_h, prepared_h],
             )?;
 
-            // eprintln!("after PC batch check: constraints: {}", cs.num_constraints());
+            if cfg!(debug_assertions) {
+                println!("after PC batch check: constraints: {}", cs.num_constraints());
+            }
 
             let rhs = &PG::GTGadget::one(cs.ns(|| "rhs"))?;
             lhs.is_eq(cs.ns(|| "lhs_is_eq_rhs"), &rhs)