--- a/polycommit/src/marlin_pc/gadgets/marlin_kzg10.rs
+++ b/polycommit/src/marlin_pc/gadgets/marlin_kzg10.rs
@@ -154,7 +154,7 @@ where
         }
 
         if cfg!(debug_assertions) {
-            println!("before PC combining commitments: constraints: {}", cs.num_constraints());
+            eprintln!("before PC combining commitments: constraints: {}", cs.num_constraints());
         }
 
         let zero = PG::G1Gadget::zero(cs.ns(|| format!("g1_zero")))?;
@@ -357,7 +357,7 @@ where
         }
 
         if cfg!(debug_assertions) {
-            println!("before PC batch check: constraints: {}", cs.num_constraints());
+            eprintln!("before PC batch check: constraints: {}", cs.num_constraints());
         }
 
         // Perform the batch check.
@@ -533,7 +533,7 @@ where
             )?;
 
             if cfg!(debug_assertions) {
-                println!("after PC batch check: constraints: {}", cs.num_constraints());
+                eprintln!("after PC batch check: constraints: {}", cs.num_constraints());
             }
 
             let rhs = &PG::GTGadget::one(cs.ns(|| "rhs"))?;