--- a/polycommit/src/sonic_pc/gadgets/sonic_kzg10.rs
+++ b/polycommit/src/sonic_pc/gadgets/sonic_kzg10.rs
@@ -180,7 +180,7 @@ where
         }
 
         if cfg!(debug_assertions) {
-            println!("before PC combining commitments: constraints: {}", cs.num_constraints());
+            eprintln!("before PC combining commitments: constraints: {}", cs.num_constraints());
         }
 
         let zero = PG::G1Gadget::zero(cs.ns(|| format!("g1_zero")))?;
@@ -347,7 +347,7 @@ where
         }
 
         if cfg!(debug_assertions) {
-            println!("before PC batch check: constraints: {}", cs.num_constraints());
+            eprintln!("before PC batch check: constraints: {}", cs.num_constraints());
         }
 
         // Perform the batch check.
@@ -554,7 +554,7 @@ where
             )?;
 
             if cfg!(debug_assertions) {
-                println!("after PC batch check: constraints: {}", cs.num_constraints());
+                eprintln!("after PC batch check: constraints: {}", cs.num_constraints());
             }
 
             let rhs = &PG::GTGadget::one(cs.ns(|| "rhs"))?;