<!-- Thank you for submitting the PR! We appreciate you spending the time to work on these changes! -->

## Motivation

This PR fixes merkle tree rebuilds and updates updates the parallelization of leaf hashing. Additionally an optimization has been added to `LedgerTree`.

## Test Plan

A rebuild test has been added to ensure that the tree is being rebuilt correctly.

## Additional Notes

Quick benchmark of the merkle tree implementations

**Old:**
- 1000 leaves:
    - new() - 500 ms
    - rebuild() - 508 ms
- 10000 leaves:
    - new() - 5768 ms
    - rebuild() - 5791 ms
- 100000 leaves:
    - new() - 52099 ms (52 seconds)
    - rebuild() - 52115 ms (52 seconds)
    
**New:**
- 1000 leaves:
    - new() - 33 ms
    - rebuild() - 280 ms
- 10000 leaves:
    - new() - 199 ms
    - rebuild() - 3667 ms
- 100000 leaves:
    - new() - 1724 ms
    - rebuild() - 31108 ms (31 seconds)